+++ src/latents/__init__.py
@@ -4,10 +4,10 @@ import contextlib
 from importlib.metadata import PackageNotFoundError, version
 
 from latents import (
+    data,
     gfa,
     mdlag,
     observation,
-    observation_model,
     plotting,
     state,
 )
@@ -16,10 +16,10 @@ with contextlib.suppress(PackageNotFoundError):
     __version__ = version("latents")
 
 __all__ = [
+    "data",
     "gfa",
     "mdlag",
     "observation",
-    "observation_model",
     "plotting",
     "state",
 ]